// components/textfield/targets/textfieldTargets.enum.ts
export enum TextfieldTargets {
  ROOT = 'root',
  CHARACTER_LIMIT = 'characterLimit',
  ERRORS = 'errors',
  HELPER_TEXT = 'helperText',
  INPUT = 'input',
  LABEL = 'label',
  PREFIX = 'prefix',
  READ_ONLY = 'readOnly',
  SIMPLE_EDIT = 'simpleEdit',
  SIMPLE_READ_ONLY = 'simpleReadOnly',
  SUFFIX = 'suffix',
  TOOLTIP_BUTTON = 'tooltipButton',
  TOOLTIP_POPOVER = 'tooltipPopover',
}